fixed barber's earnings field type

+++ PRPA/Models/Barber.cs
@@ -3,7 +3,7 @@
     public class Barber
     {
         public int BarberId { get; set; }
-        public int Earnings { get; set; }
+        public decimal Earnings { get; set; }
         public int UserID { get; set; }
     }
 }
